GUI.
Refactored the creation of pizzas to move any duplicate code into the PizzaStore super class.

Added some functionalities to model classes to allow attributes to be exposed through the GUI. These include adding a name to each handler, adding a method to add a single pizza to an order in an order class, adding a method which returns the pizzas which are in an order. Added a method to the PizzaStore class which exposes the names of pizza on the menu. Added a hashmap which relates the name to be displayed in the GUI to the key used to create the pizzas.

Implemented an MVC pattern.
Added three views: Shop front, store front and order. Added a model class which has order, customer and selected store as attribute. Added a controller which adds action listeners to the view elements which manipulate the model class.